# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tako Master Server Control Plane API
  version: 1.0.0
  description: |
    REST API specification for Tako Master Server.
    Provides cluster orchestration, node management, application lifecycles, and deployment pipelines.
servers:
  - url: http://localhost:8080
    description: Local Master Server
  - url: https://api.gettako.dev
    description: Production Master Server
paths:
...
  /api/v1/services:
    get:
      summary: List all services
      operationId: listServices
      tags: [Services]
      responses:
        '200':
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
    post:
      summary: Create service
      operationId: createService
      tags: [Services]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateServiceInput'
      responses:
        '201':
          description: Service created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
  /api/v1/services/{id}:
    get:
      summary: Get service by ID
      operationId: getService
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '404':
          description: Service not found
    delete:
      summary: Delete service
      operationId: deleteService
      tags: [Services]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Service deleted
  /api/v1/services/{id}/deploy:
    post:
      summary: Trigger deployment pipeline for service
      operationId: triggerDeploy
      tags: [Services]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Deployment accepted and queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  deploymentId:
                    type: string
                    example: dep-9a8b7c6d
                  status:
                    type: string
                    example: queued
  /api/v1/services/{id}/deployments:
    get:
      summary: List deployments for a service
      operationId: listServiceDeployments
      tags: [Services]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployments history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
  /api/v1/deployments:
    get:
      summary: Query deployments
      operationId: listDeployments
      tags: [Deployments]
      parameters:
        - name: serviceId
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Deployment'
  /api/v1/deployments/{id}:
    get:
      summary: Get deployment details
      operationId: getDeployment
      tags: [Deployments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deployment details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Deployment'
  /api/v1/deployments/{id}/logs:
    get:
      summary: Stream live deployment build logs (SSE)
      operationId: streamDeploymentLogs
      tags: [Deployments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Realtime log chunks stream
          content:
            text/event-stream:
              schema:
                type: string
  /api/v1/deployments/{id}/logs/history:
    get:
      summary: Get full historical deployment logs as plain text
      operationId: getDeploymentLogsHistory
      tags: [Deployments]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Plaintext log history
          content:
            text/plain:
              schema:
                type: string
...
  /api/v1/settings:
    get:
      summary: Get all cluster settings
      operationId: getClusterSettings
      tags: [Settings]
      responses:
        '200':
          description: Key-value map of cluster settings
          content:
            application/json:
              schema:
                type: object
    post:
      summary: Batch update cluster settings
      operationId: updateClusterSettingsBatch
      tags: [Settings]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Settings successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
  /api/v1/settings/{key}:
    get:
      summary: Get cluster setting by key
      operationId: getClusterSettingByKey
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Setting value
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Setting not found
    put:
      summary: Update or create cluster setting by key
      operationId: updateClusterSettingByKey
      tags: [Settings]
      parameters:
        - name: key
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [value]
              properties:
                value:
                  type: object
      responses:
        '200':
          description: Setting updated
          content:
            application/json:
              schema:
                type: object
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateServiceInput:
      type: object
      required: [projectId, nodeId, name, type]
      properties:
        projectId:
          type: string
        nodeId:
          type: string
        name:
          type: string
        slug:
          type: string
        type:
          type: string
          enum: [app, compose, database]
        repository:
          type: string
        branch:
          type: string
        dockerfile:
          type: string
        image:
          type: string
        ports:
          type: array
          items:
            type: integer
        domains:
          type: array
          items:
            type: string
        environmentVars:
          type: object
          additionalProperties:
            type: string
    Service:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
        nodeId:
          type: string
        name:
          type: string
        slug:
          type: string
        type:
          type: string
        status:
          type: string
        repository:
          type: string
        branch:
          type: string
        image:
          type: string
        ports:
          type: array
          items:
            type: integer
        domains:
          type: array
          items:
            type: string
        createdAt:
          type: string
        updatedAt:
          type: string
    Deployment:
      type: object
      properties:
        id:
          type: string
        serviceId:
          type: string
        commitHash:
          type: string
        commitMessage:
          type: string
        branch:
          type: string
        author:
          type: string
        status:
          type: string
        steps:
          type: string
        logs:
          type: string
        createdAt:
          type: string
        finishedAt:
          type: string
        durationMs:
          type: integer
...